+++ packages/flame/lib/src/effects/move_effect.dart
@@ -1,139 +1,98 @@
 import 'dart:ui';
 
-import 'package:flutter/animation.dart';
+import 'package:vector_math/vector_math_64.dart';
 
-import '../../components.dart';
-import '../extensions/vector2.dart';
-import 'effects.dart';
+import 'effect_controller.dart';
+import 'transform2d_effect.dart';
 
-class Vector2Percentage {
-  final Vector2 v;
-  final Vector2 previous;
-  final double startAt;
-  final double endAt;
+/// Move a component to a new position.
+///
+/// The following constructors are provided:
+///
+///   - [MoveEffect.by] will move the target in a straight line to a new
+///     position that is at an `offset` from the target's position at the onset
+///     of the effect;
+///
+///   - [MoveEffect.to] will move the target in a straight line to the specified
+///     coordinates;
+///
+///   - [MoveEffect.along] will move the target along the specified path, which
+///     may contain curved segments, but must be simply-connected. The `path`
+///     argument is taken as relative to the target's position at the start of
+///     the effect.
+///
+/// This effect applies incremental changes to the component's position, and
+/// requires that any other effect or update logic applied to the same component
+/// also used incremental updates.
+class MoveEffect extends Transform2DEffect {
+  MoveEffect.by(Vector2 offset, EffectController controller)
+      : _offset = offset.clone(),
+        super(controller);
 
-  Vector2Percentage(
-    this.v,
-    this.previous,
-    this.startAt,
-    this.endAt,
-  );
+  factory MoveEffect.to(Vector2 destination, EffectController controller) =>
+      _MoveToEffect(destination, controller);
+
+  factory MoveEffect.along(Path path, EffectController controller) =>
+      _MoveAlongPathEffect(path, controller);
+
+  Vector2 _offset;
+
+  @override
+  void apply(double progress) {
+    final dProgress = progress - previousProgress;
+    target.position += _offset * dProgress;
+    super.apply(progress);
+  }
 }
 
-class MoveEffect extends PositionComponentEffect {
-  List<Vector2> path;
-  Vector2Percentage? _currentSubPath;
-  List<Vector2Percentage>? _percentagePath;
+/// Implementation class for [MoveEffect.to]
+class _MoveToEffect extends MoveEffect {
+  _MoveToEffect(Vector2 destination, EffectController controller)
+      : _destination = destination.clone(),
+        super.by(Vector2.zero(), controller);
 
-  /// Duration or speed needs to be defined
-  MoveEffect({
-    required this.path,
-    double? duration,
-    double? speed,
-    Curve? curve,
-    bool isInfinite = false,
-    bool isAlternating = false,
-    bool isRelative = false,
-    double? initialDelay,
-    double? peakDelay,
-    bool? removeOnFinish,
-    VoidCallback? onComplete,
-  }) : super(
-          isInfinite,
-          isAlternating,
-          duration: duration,
-          speed: speed,
-          curve: curve,
-          isRelative: isRelative,
-          modifiesPosition: true,
-          initialDelay: initialDelay,
-          peakDelay: peakDelay,
-          removeOnFinish: removeOnFinish,
-          onComplete: onComplete,
-        );
+  final Vector2 _destination;
 
   @override
-  Future<void> onLoad() async {
-    super.onLoad();
-    List<Vector2> _movePath;
-    // With relative here we mean that any vector in the list is relative
-    // to the previous vector in the list, except the first one which is
-    // relative to the start position of the component.
-    if (isRelative) {
-      var lastPosition = originalPosition!;
-      _movePath = [];
-      for (final v in path) {
-        final nextPosition = v + lastPosition;
-        _movePath.add(nextPosition);
-        lastPosition = nextPosition;
-      }
-    } else {
-      _movePath = path;
-    }
-    peakPosition = _movePath.last;
+  void onStart() {
+    _offset = _destination - target.position;
+  }
+}
 
-    var pathLength = 0.0;
-    var lastPosition = originalPosition!;
-    for (final v in _movePath) {
-      pathLength += v.distanceTo(lastPosition);
-      lastPosition = v;
-    }
-
-    _percentagePath = <Vector2Percentage>[];
-    lastPosition = originalPosition!;
-    for (final v in _movePath) {
-      final lengthToPrevious = lastPosition.distanceTo(v);
-      final lastEndAt =
-          _percentagePath!.isNotEmpty ? _percentagePath!.last.endAt : 0.0;
-      final endPercentage = lastEndAt + lengthToPrevious / pathLength;
-      _percentagePath!.add(
-        Vector2Percentage(
-          v,
-          lastPosition,
-          lastEndAt,
-          _movePath.last == v ? 1.0 : endPercentage,
-        ),
+/// Implementation class for [MoveEffect.along].
+class _MoveAlongPathEffect extends MoveEffect {
+  _MoveAlongPathEffect(Path path, EffectController controller)
+      : _lastPosition = Vector2.zero(),
+        super.by(Vector2.zero(), controller) {
+    final metrics = path.computeMetrics().toList();
+    if (metrics.length != 1) {
+      throw ArgumentError(
+        'Only single-contour paths are allowed in MoveEffect.along',
       );
-      lastPosition = v;
     }
-    final totalPathLength = isAlternating ? pathLength * 2 : pathLength;
-    speed ??= totalPathLength / duration!;
+    _pathMetric = metrics[0];
+    _pathLength = _pathMetric.length;
+    assert(_pathLength > 0);
+  }
 
-    // `duration` is not null when speed is null
-    duration ??= totalPathLength / speed!;
-    duration = duration! + totalOffset;
-    setPeakTimeFromDuration(duration!);
+  late final PathMetric _pathMetric;
+  late final double _pathLength;
+  late Vector2 _lastPosition;
+
+  @override
+  void apply(double progress) {
+    final distance = progress * _pathLength;
+    final tangent = _pathMetric.getTangentForOffset(distance)!;
+    final offset = tangent.position;
+    target.position.x += offset.dx - _lastPosition.x;
+    target.position.y += offset.dy - _lastPosition.y;
+    _lastPosition.setValues(offset.dx, offset.dy);
+    super.apply(progress);
   }
 
   @override
-  void resetEffect() {
-    super.resetEffect();
-    if (_percentagePath?.isNotEmpty ?? false) {
-      _currentSubPath = _percentagePath!.first;
-    }
-  }
-
-  @override
-  void update(double dt) {
-    if (isPaused) {
-      return;
-    }
-    super.update(dt);
-    if (hasCompleted()) {
-      return;
-    }
-    _currentSubPath ??= _percentagePath!.first;
-    if (!curveDirection.isNegative && _currentSubPath!.endAt < curveProgress ||
-        curveDirection.isNegative && _currentSubPath!.startAt > curveProgress) {
-      _currentSubPath =
-          _percentagePath!.firstWhere((v) => v.endAt >= curveProgress);
-    }
-    final lastEndAt = _currentSubPath!.startAt;
-    final localPercentage =
-        (curveProgress - lastEndAt) / (_currentSubPath!.endAt - lastEndAt);
-    affectedParent.position.setFrom(
-      _currentSubPath!.previous +
-          ((_currentSubPath!.v - _currentSubPath!.previous) * localPercentage),
-    );
+  void reset() {
+    super.reset();
+    _lastPosition = Vector2.zero();
   }
 }